backend/requirements.txt
# Web framework
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.32.0

# Database
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
alembic>=1.14.0

# Auth
pyjwt>=2.10.0
bcrypt>=4.2.0
python-multipart>=0.0.17

# Google Calendar
google-api-python-client>=2.155.0
google-auth>=2.36.0

# Validation & Utils
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-dateutil>=2.9.0

# Dev / Test
pytest>=8.3.0
pytest-asyncio>=0.24.0
httpx>=0.28.0
